requirements.txt
python-pptx>=0.6.23
gspread>=6.0.0
google-auth>=2.0.0
matplotlib>=3.8.0
streamlit>=1.30.0
Pillow>=10.0.0
